+++ example/native/hub/src/sample_functions.rs
@@ -101,7 +101,9 @@ pub async fn stream_mandelbrot() {
 
 pub async fn run_debug_tests() {
     crate::sleep(std::time::Duration::from_secs(1)).await;
-    crate::debug_print!("\nStarting debug tests.");
+    crate::debug_print!("Starting debug tests.");
+    let timestamp = sample_crate::get_current_time();
+    crate::debug_print!("System time is {}", timestamp);
     let option = sample_crate::compilation_test::get_hardward_id();
     if let Some(hwid) = option {
         crate::debug_print!("Hardware ID is {}.", hwid);
@@ -118,5 +120,5 @@ pub async fn run_debug_tests() {
             crate::debug_print!("Counted to {}, yielding regularly.", count);
         }
     }
-    crate::debug_print!("Debug tests completed!\n");
+    crate::debug_print!("Debug tests completed!");
 }
