+++ example/native/sample_crate/Cargo.toml
@@ -11,3 +11,4 @@ anyhow = "1.0.69"
 num = "0.4"
 image = "0.24.3"
 crossbeam = "0.8"
+chrono = { version = "0.4.31", features = ["wasmbind"] }
